+++ src/main/java/net/canadensys/dataportal/occurrence/model/ModelFieldConstants.java
@@ -1,0 +1,12 @@
+package net.canadensys.dataportal.occurrence.model;
+
+/**
+ * Constants definitions for model fields(rows) name.
+ * @author cgendreau
+ *
+ */
+public class ModelFieldConstants {
+	
+	public static final String SOURCE_FILE_ID = "sourcefileid";
+
+}