+++ src/main/resources/script/occurrence/create_occurrence_tables.sql
@@ -1,8 +1,7 @@
 CREATE TABLE occurrence (
 auto_id INTEGER NOT NULL,
 associatedmedia TEXT,
-associatedsequences TEXT,
-basisofrecord VARCHAR(25),
+associatedmediamime TEXT,
 catalogNumber TEXT,
 _references TEXT,
 collectionCode TEXT,
@@ -274,6 +273,18 @@ CREATE TABLE resource_contact
 	phone character varying(20),
 	email character varying(200),
 	CONSTRAINT resource_contact_pkey PRIMARY KEY (id)
+);
+
+CREATE SEQUENCE resource_management_id_seq;
+CREATE TABLE resource_management
+(
+  id integer DEFAULT nextval('resource_management_id_seq') NOT NULL,
+  sourcefileid character varying(50),
+  name character varying(255),
+  key character varying(36),
+  archive_url character varying(255),
+  CONSTRAINT resource_management_pkey PRIMARY KEY ( id ),
+  CONSTRAINT resource_management_source_file_id_key UNIQUE (sourcefileid)
 );
 
 CREATE OR REPLACE VIEW complete_occurrence_view AS